Move MockWebRTCAdaptor to under test directory
