+++ src/test/java/io/antmedia/test/webrtc/adaptor/MockWebRTCAdaptor.java
@@ -1,4 +1,4 @@
-package io.antmedia.webrtc;
+package io.antmedia.test.webrtc.adaptor;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -7,6 +7,7 @@ import java.util.Set;
 
 import io.antmedia.cluster.IStreamInfo;
 import io.antmedia.rest.WebRTCClientStats;
+import io.antmedia.webrtc.VideoCodec;
 import io.antmedia.webrtc.api.IWebRTCAdaptor;
 import io.antmedia.webrtc.api.IWebRTCClient;
 import io.antmedia.webrtc.api.IWebRTCMuxer;
